kafkaproducer/requirements.txt
kafka-python
requests
python-dateutil